Diff and import songs from another music archive using the DB
This matches songs between the Pony Music Archive and another
collection that has been converted to a PMA-compatible layout.
After creating a DB for both archives we can diff their contents,
and keep the best of the two for each matching song,
then import the rest of the songs we don't already have.

Start with a high threshold for the Chromaprint matcher,
then reduce progressively while watching for any false positives.
Going below an 80% match is typically not worth it, there'll be
too many false positives to sort through and few legitimate matches.