Automatically find and import higher-quality MP3s
We reject files with a sketchy audio spectrum, those files should
be inspected manually as they are sometimes actually slightly better.

Most of the time what gets flagged is slightly larger files with a
slighly higher bitrate, but 48kHz MP3 instead of 44.1kHz, so the
spectrum ends up looking a little bit worse in the end.
It's a manual decision whether to import these files or not.